kanvas/requirements.txt
django==1.8.5
django-tinymce==2.0.6
django-allauth==0.24.1
sorl-thumbnail==12.3

